Moved back to single header and implementation
