--- MBBackgroundView.h
@@ -1,15 +1,0 @@
-//
-//  MBBackgroundView.h
-//  MBProgressHUD
-//
-//  Created by Goff Marocchi on 05/03/2017.
-//  Copyright © 2017 Matej Bukovinski. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-#import "ProgressBackgroundViewing.h"
-
-@interface MBBackgroundView : UIView <ProgressBackgroundViewing>
-
-@end
