--- MBProgressHUDRoundedButton.h
@@ -1,13 +1,0 @@
-//
-//  MBProgressHUDRoundedButton.h
-//  MBProgressHUD
-//
-//  Created by Goff Marocchi on 05/03/2017.
-//  Copyright © 2017 Matej Bukovinski. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-@interface MBProgressHUDRoundedButton : UIButton
-
-@end
