--- MBProgressHUDRoundedButton.m
@@ -1,57 +1,0 @@
-//
-//  MBProgressHUDRoundedButton.m
-//  MBProgressHUD
-//
-//  Created by Goff Marocchi on 05/03/2017.
-//  Copyright © 2017 Matej Bukovinski. All rights reserved.
-//
-
-#import "MBProgressHUDRoundedButton.h"
-
-@implementation MBProgressHUDRoundedButton
-
-#pragma mark - Lifecycle
-
-- (instancetype)initWithFrame:(CGRect)frame {
-    self = [super initWithFrame:frame];
-    if (self) {
-        CALayer *layer = self.layer;
-        layer.borderWidth = 1.f;
-    }
-    return self;
-}
-
-#pragma mark - Layout
-
-- (void)layoutSubviews {
-    [super layoutSubviews];
-    // Fully rounded corners
-    CGFloat height = CGRectGetHeight(self.bounds);
-    self.layer.cornerRadius = ceil(height / 2.f);
-}
-
-- (CGSize)intrinsicContentSize {
-    // Only show if we have associated control events
-    if (self.allControlEvents == 0) return CGSizeZero;
-    CGSize size = [super intrinsicContentSize];
-    // Add some side padding
-    size.width += 20.f;
-    return size;
-}
-
-#pragma mark - Color
-
-- (void)setTitleColor:(UIColor *)color forState:(UIControlState)state {
-    [super setTitleColor:color forState:state];
-    // Update related colors
-    [self setHighlighted:self.highlighted];
-    self.layer.borderColor = color.CGColor;
-}
-
-- (void)setHighlighted:(BOOL)highlighted {
-    [super setHighlighted:highlighted];
-    UIColor *baseColor = [self titleColorForState:UIControlStateSelected];
-    self.backgroundColor = highlighted ? [baseColor colorWithAlphaComponent:0.1f] : [UIColor clearColor];
-}
-
-@end
