--- MBRoundProgressView.h
@@ -1,39 +1,0 @@
-//
-//  MBRoundProgressView.h
-//  MBProgressHUD
-//
-//  Created by Goff Marocchi on 05/03/2017.
-//  Copyright © 2017 Matej Bukovinski. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-/**
- * A progress view for showing definite progress by filling up a circle (pie chart).
- */
-@interface MBRoundProgressView : UIView
-
-/**
- * Progress (0.0 to 1.0)
- */
-@property (nonatomic, assign) float progress;
-
-/**
- * Indicator progress color.
- * Defaults to white [UIColor whiteColor].
- */
-@property (nonatomic, strong) UIColor *progressTintColor;
-
-/**
- * Indicator background (non-progress) color.
- * Only applicable on iOS versions older than iOS 7.
- * Defaults to translucent white (alpha 0.1).
- */
-@property (nonatomic, strong) UIColor *backgroundTintColor;
-
-/*
- * Display mode - NO = round or YES = annular. Defaults to round.
- */
-@property (nonatomic, assign, getter = isAnnular) BOOL annular;
-
-@end
