--- MBRoundProgressView.m
@@ -1,129 +1,0 @@
-//
-//  MBRoundProgressView.m
-//  MBProgressHUD
-//
-//  Created by Goff Marocchi on 05/03/2017.
-//  Copyright © 2017 Matej Bukovinski. All rights reserved.
-//
-
-#import "MBRoundProgressView.h"
-
-@implementation MBRoundProgressView
-
-#pragma mark - Lifecycle
-
-- (id)init {
-    return [self initWithFrame:CGRectMake(0.f, 0.f, 37.f, 37.f)];
-}
-
-- (id)initWithFrame:(CGRect)frame {
-    self = [super initWithFrame:frame];
-    if (self) {
-        self.backgroundColor = [UIColor clearColor];
-        self.opaque = NO;
-        _progress = 0.f;
-        _annular = NO;
-        _progressTintColor = [[UIColor alloc] initWithWhite:1.f alpha:1.f];
-        _backgroundTintColor = [[UIColor alloc] initWithWhite:1.f alpha:.1f];
-    }
-    return self;
-}
-
-#pragma mark - Layout
-
-- (CGSize)intrinsicContentSize {
-    return CGSizeMake(37.f, 37.f);
-}
-
-#pragma mark - Properties
-
-- (void)setProgress:(float)progress {
-    if (progress != _progress) {
-        _progress = progress;
-        [self setNeedsDisplay];
-    }
-}
-
-- (void)setProgressTintColor:(UIColor *)progressTintColor {
-    NSAssert(progressTintColor, @"The color should not be nil.");
-    if (progressTintColor != _progressTintColor && ![progressTintColor isEqual:_progressTintColor]) {
-        _progressTintColor = progressTintColor;
-        [self setNeedsDisplay];
-    }
-}
-
-- (void)setBackgroundTintColor:(UIColor *)backgroundTintColor {
-    NSAssert(backgroundTintColor, @"The color should not be nil.");
-    if (backgroundTintColor != _backgroundTintColor && ![backgroundTintColor isEqual:_backgroundTintColor]) {
-        _backgroundTintColor = backgroundTintColor;
-        [self setNeedsDisplay];
-    }
-}
-
-#pragma mark - Drawing
-
-- (void)drawRect:(CGRect)rect {
-    CGContextRef context = UIGraphicsGetCurrentContext();
-    BOOL isPreiOS7 = kCFCoreFoundationVersionNumber < kCFCoreFoundationVersionNumber_iOS_7_0;
-    
-    if (_annular) {
-        // Draw background
-        CGFloat lineWidth = isPreiOS7 ? 5.f : 2.f;
-        UIBezierPath *processBackgroundPath = [UIBezierPath bezierPath];
-        processBackgroundPath.lineWidth = lineWidth;
-        processBackgroundPath.lineCapStyle = kCGLineCapButt;
-        CGPoint center = CGPointMake(CGRectGetMidX(self.bounds), CGRectGetMidY(self.bounds));
-        CGFloat radius = (self.bounds.size.width - lineWidth)/2;
-        CGFloat startAngle = - ((float)M_PI / 2); // 90 degrees
-        CGFloat endAngle = (2 * (float)M_PI) + startAngle;
-        [processBackgroundPath addArcWithCenter:center radius:radius startAngle:startAngle endAngle:endAngle clockwise:YES];
-        [_backgroundTintColor set];
-        [processBackgroundPath stroke];
-        // Draw progress
-        UIBezierPath *processPath = [UIBezierPath bezierPath];
-        processPath.lineCapStyle = isPreiOS7 ? kCGLineCapRound : kCGLineCapSquare;
-        processPath.lineWidth = lineWidth;
-        endAngle = (self.progress * 2 * (float)M_PI) + startAngle;
-        [processPath addArcWithCenter:center radius:radius startAngle:startAngle endAngle:endAngle clockwise:YES];
-        [_progressTintColor set];
-        [processPath stroke];
-    } else {
-        // Draw background
-        CGFloat lineWidth = 2.f;
-        CGRect allRect = self.bounds;
-        CGRect circleRect = CGRectInset(allRect, lineWidth/2.f, lineWidth/2.f);
-        CGPoint center = CGPointMake(CGRectGetMidX(self.bounds), CGRectGetMidY(self.bounds));
-        [_progressTintColor setStroke];
-        [_backgroundTintColor setFill];
-        CGContextSetLineWidth(context, lineWidth);
-        if (isPreiOS7) {
-            CGContextFillEllipseInRect(context, circleRect);
-        }
-        CGContextStrokeEllipseInRect(context, circleRect);
-        // 90 degrees
-        CGFloat startAngle = - ((float)M_PI / 2.f);
-        // Draw progress
-        if (isPreiOS7) {
-            CGFloat radius = (CGRectGetWidth(self.bounds) / 2.f) - lineWidth;
-            CGFloat endAngle = (self.progress * 2.f * (float)M_PI) + startAngle;
-            [_progressTintColor setFill];
-            CGContextMoveToPoint(context, center.x, center.y);
-            CGContextAddArc(context, center.x, center.y, radius, startAngle, endAngle, 0);
-            CGContextClosePath(context);
-            CGContextFillPath(context);
-        } else {
-            UIBezierPath *processPath = [UIBezierPath bezierPath];
-            processPath.lineCapStyle = kCGLineCapButt;
-            processPath.lineWidth = lineWidth * 2.f;
-            CGFloat radius = (CGRectGetWidth(self.bounds) / 2.f) - (processPath.lineWidth / 2.f);
-            CGFloat endAngle = (self.progress * 2.f * (float)M_PI) + startAngle;
-            [processPath addArcWithCenter:center radius:radius startAngle:startAngle endAngle:endAngle clockwise:YES];
-            // Ensure that we don't get color overlaping when _progressTintColor alpha < 1.f.
-            CGContextSetBlendMode(context, kCGBlendModeCopy);
-            [_progressTintColor set];
-            [processPath stroke];
-        }
-    }
-}
-
-@end
